requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
transformers
types-python-jose
types-requests
passlib[bcrypt]
python-jose
tenacity
pymysql
cryptography